requirements.txt
google-generativeai
python-dotenv
fastapi
uvicorn
sse-starlette
pydantic
google-auth
google-auth-oauthlib
google-auth-httplib2
google-cloud-storage
pyjwt
requests
motor
rapidfuzz
markdown
httpx
